chore: load form theme options
