+++ src/Form/Theme.php
@@ -9,10 +9,6 @@
 
 namespace Give\Form;
 
-use Give\Form\Theme\ThemeOptions;
-use InvalidArgumentException;
-use function Give\Helpers\Form\Theme\get as getTheme;
-
 defined( 'ABSPATH' ) || exit;
 
 /**
@@ -53,63 +49,7 @@ abstract class Theme {
 	 *
 	 * @since 2.7.0
 	 *
-	 * @return string
+	 * @return array
 	 */
-	abstract public function getOptionsConfig();
-
-	/**
-	 * return theme options.
-	 *
-	 * @return ThemeOptions
-	 * @since 2.7.0
-	 */
-	public function getOptions() {
-		return new ThemeOptions( $this->getOptionsConfig() );
-	}
-
-	/**
-	 * return theme options.
-	 *
-	 * @since 2.7.0
-	 *
-	 * @return string
-	 */
-	public function renderOptions() {
-		global $post;
-		ob_start();
-
-		$saveOptions = getTheme( $post->ID, $this->getID() );
-
-		foreach ( $this->getOptions() as $groupID => $option ) {
-			printf(
-				'<div class="give-row %1$s">',
-				$groupID
-			);
-
-			printf(
-				'<div class="give-row-head">
-							<button type="button" class="handlediv" aria-expanded="true">
-								<span class="toggle-indicator"/>
-							</button>
-							<h2 class="hndle"><span>%1$s</span></h2>
-						</div>',
-				$option['name']
-			);
-
-			echo '<div class="give-row-body">';
-			foreach ( $option['fields'] as $field ) {
-				if ( isset( $saveOptions[ $groupID ][ $field['id'] ] ) ) {
-					$field['attributes']['value'] = $saveOptions[ $groupID ][ $field['id'] ];
-				}
-
-				$field['id'] = "{$this->getID()}[{$groupID}][{$field['id']}]";
-
-				give_render_field( $field );
-			}
-
-			echo '</div></div>';
-		}
-
-		return ob_get_clean();
-	}
+	abstract public function getOptions();
 }
