--- src/Form/ThemeOptions.php
@@ -1,20 +1,0 @@
-<?php
-namespace Give\Form\Theme;
-
-/**
- * Class Options
- *
- * @since 2.7.0
- * @package Give\Form\Theme
- */
-class ThemeOptions {
-
-	/**
-	 * ThemeOptions constructor.
-	 *
-	 * @since 2.7.0
-	 * @param array $args
-	 */
-	public function __construct( $args ) {
-	}
-}
